requirements.txt
Flask
flask-cors
numpy
scikit-learn
pandas
matplotlib
python-dotenv
gensim